Push does not return the new size (#198)
* push(x) does not return the new size.

* Updated failing execution tests containing push.

* Updated failing compilation tests.

+++ test/libsolidity/semanticTests/viaYul/array_storage_length_access.sol
@@ -1,6 +1,6 @@
 contract C {
     uint[] storageArray;
-    function set_get_length(uint256 len) public returns (uint256) {
+    function set_get_length(uint len) public returns (uint) {
         while(storageArray.length < len)
             storageArray.push();
         return storageArray.length;
@@ -9,10 +9,10 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// set_get_length(uint256): 0 -> 0
-// set_get_length(uint256): 1 -> 1
-// set_get_length(uint256): 10 -> 10
-// set_get_length(uint256): 20 -> 20
-// set_get_length(uint256): 0xFF -> 0xFF
-// set_get_length(uint256): 0xFFF -> 0xFFF
-// set_get_length(uint256): 0xFFFFF -> FAILURE # Out-of-gas #
+// set_get_length(uint): 0 -> 0
+// set_get_length(uint): 1 -> 1
+// set_get_length(uint): 10 -> 10
+// set_get_length(uint): 20 -> 20
+// set_get_length(uint): 0x00FF -> 0x00FF
+// set_get_length(uint): 0x0FFF -> 0x0FFF
+// set_get_length(uint): 0x0FFFFF -> FAILURE, 5
